% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sncf.R
\name{get_sncf}
\alias{get_sncf}
\title{scrap sncf}
\usage{
get_sncf(dataset = "ponctualite-mensuelle-transilien")
}
\arguments{
\item{dataset}{a character}
}
\value{
a json document\code{data.frame}
}
\description{
récupère des données sur le site sncf.com
}
\examples{
\dontrun{
  get_sncf("ponctualite-mensuelle-transilien")
}

}
