% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Moyenne et médiane sans valeur manquantes}
\usage{
impute_na(x, method = c("moyenne", "mediane"))
}
\arguments{
\item{x}{a vector}

\item{method}{a choice of "médiane" or "mean"}
}
\value{
a vector
}
\description{
Remplace les valeurs manquantes par la médiane ou la moyenne.
}
\examples{
\dontrun{
 impute_na(x, "moyenne")
}


}
